feat: renames popular people app bar widget file to match widget name
